+++ src/validate.ts
@@ -2,9 +2,67 @@
 // Distributed under the terms of the Modified BSD License.
 'use strict';
 
-import { IKernelId, IKernelSpecId } from './ikernel';
+import { 
+  IKernelId, IKernelMessage, IKernelMessageHeader, IKernelSpecId
+} from './ikernel';
 
 import { INotebookId, ISessionId } from './isession';
+
+
+/**
+ * Validate an Kernel Message as being a valid Comm Message.
+ */
+export 
+function validateCommMessage(msg: IKernelMessage): boolean {
+  var fields = ['comm_id', 'target_name'];
+  for (var i = 0; i < fields.length; i++) {
+    if (!msg.content.hasOwnProperty(fields[i])) {
+      return false;
+    }
+    if (typeof msg.content[fields[i]] !== 'string') {
+      return false;
+    }
+  } 
+  return true;
+}
+
+
+function validateKernelHeader(header: any): void {
+  var fields = ['username', 'version', 'session', 'msg_id',  'msg_type'];
+  for (var i = 0; i < fields.length; i++) {
+    if (!header.hasOwnProperty(fields[i])) {
+      throw Error('Invalid Kernel message ' + fields[i]);
+    }
+    if (typeof header[fields[i]] !== 'string') {
+      throw Error('Invalid Kernel message');
+    }
+  }
+}
+
+
+/**
+ * Validate an object as being of IKernelMessage type.
+ */
+export
+function validateKernelMessage(msg: IKernelMessage) : void {
+  var fields = ['header', 'parent_header', 'metadata', 'content', 
+                'channel', 'buffers'];
+  for (var i = 0; i < fields.length; i++) {
+    if (!msg.hasOwnProperty(fields[i])) {
+      throw Error('Invalid Kernel message');
+    }
+  }
+  validateKernelHeader(msg.header);
+  if (Object.keys(msg.parent_header).length > 0) {
+    validateKernelHeader(msg.parent_header as IKernelMessageHeader);
+  }
+  if (typeof msg.channel !== 'string') {
+    throw Error('Invalid Kernel message');
+  }
+  if (!Array.isArray(msg.buffers)) {
+    throw Error('Invalid Kernel message');
+  }
+}
 
 
 /**
